chore: Unify chat overlay views (out of office & pinned messages)
