+++ NextcloudTalk/Chat/Chat views/ChatOverlayView.xib
@@ -1,22 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="24128" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="24506" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
     <device id="retina6_12" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="24063"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="24504"/>
         <capability name="System colors in document resources" minToolsVersion="11.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="OutOfOfficeView" customModule="NextcloudTalk" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ChatOverlayView" customModule="NextcloudTalk" customModuleProvider="target">
             <connections>
                 <outlet property="backgroundView" destination="Xsl-mz-DBd" id="gvK-dc-O5I"/>
                 <outlet property="contentView" destination="iN0-l3-epB" id="gDR-On-leO"/>
-                <outlet property="dates" destination="0sH-vg-JfJ" id="nmq-2I-Q4r"/>
                 <outlet property="leftIndicator" destination="MdQ-4Z-aXZ" id="Op4-Ve-fJg"/>
-                <outlet property="replacement" destination="335-sd-Dif" id="2qP-rw-hYL"/>
+                <outlet property="secondarySubtitle" destination="335-sd-Dif" id="2qP-rw-hYL"/>
                 <outlet property="stackView" destination="OHJ-nY-k0O" id="X43-BC-ZYX"/>
-                <outlet property="subtitle" destination="fGe-4e-Cfj" id="iKQ-Vp-YZq"/>
+                <outlet property="subtitle" destination="0sH-vg-JfJ" id="nmq-2I-Q4r"/>
+                <outlet property="textView" destination="fGe-4e-Cfj" id="iKQ-Vp-YZq"/>
                 <outlet property="title" destination="bqD-D9-nxa" id="ENc-AK-hlt"/>
                 <outlet property="uiMenuButton" destination="7a4-gY-YTH" id="t2f-g3-rKw"/>
                 <outlet property="wrapperView" destination="NAY-L9-7WT" id="jlg-oC-Yah"/>
@@ -28,7 +28,7 @@
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
                 <button opaque="NO" alpha="0.0" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="7a4-gY-YTH" userLabel="UIMenu Button">
-                    <rect key="frame" x="0.0" y="259" width="592" height="20"/>
+                    <rect key="frame" x="0.0" y="271" width="592" height="20"/>
                     <constraints>
                         <constraint firstAttribute="height" constant="20" id="frJ-v3-j3R"/>
                     </constraints>
@@ -71,12 +71,13 @@
                                     <nil key="textColor"/>
                                     <nil key="highlightedColor"/>
                                 </label>
-                                <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" scrollEnabled="NO" editable="NO" textAlignment="natural" selectable="NO" translatesAutoresizingMaskIntoConstraints="NO" id="fGe-4e-Cfj">
+                                <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" scrollEnabled="NO" editable="NO" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="fGe-4e-Cfj" customClass="MessageBodyTextView">
                                     <rect key="frame" x="0.0" y="85" width="546" height="166"/>
                                     <string key="text">Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.</string>
                                     <color key="textColor" systemColor="labelColor"/>
                                     <fontDescription key="fontDescription" style="UICTFontTextStyleBody"/>
                                     <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
+                                    <dataDetectorType key="dataDetectorTypes" link="YES"/>
                                 </textView>
                             </subviews>
                         </stackView>
@@ -102,7 +103,7 @@
                 <constraint firstAttribute="trailing" secondItem="7a4-gY-YTH" secondAttribute="trailing" id="16s-TC-fK2"/>
                 <constraint firstItem="NAY-L9-7WT" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="8" id="HnT-UZ-9eX"/>
                 <constraint firstItem="NAY-L9-7WT" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="8" id="lft-oK-hqM"/>
-                <constraint firstAttribute="bottom" secondItem="7a4-gY-YTH" secondAttribute="bottom" constant="12" id="mH5-06-Z5X"/>
+                <constraint firstAttribute="bottom" secondItem="7a4-gY-YTH" secondAttribute="bottom" id="mH5-06-Z5X"/>
                 <constraint firstItem="7a4-gY-YTH" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="mqW-xd-c9H"/>
                 <constraint firstAttribute="bottom" secondItem="NAY-L9-7WT" secondAttribute="bottom" constant="8" id="n1k-OH-DIC"/>
                 <constraint firstAttribute="trailing" secondItem="NAY-L9-7WT" secondAttribute="trailing" constant="8" id="nZl-aZ-jwM"/>
@@ -119,7 +120,7 @@
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
         </systemColor>
         <systemColor name="systemBlueColor">
-            <color red="0.0" green="0.47843137250000001" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <color red="0.0" green="0.53333333333333333" blue="1" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </systemColor>
     </resources>
 </document>
